+++ ggconfigd/bin/ggconfigd.c
@@ -3,19 +3,56 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "ggconfigd.h"
+#include <argp.h>
+#include <ggl/buffer.h>
+#include <ggl/object.h>
 #include <stdlib.h>
+
+static char doc[] = "ggconfigd -- Greengrass Lite configuration daemon";
+
+static struct argp_option opts[] = {
+    { "config-file", 'c', "path", 0, "Configuration file to use", 0 },
+    { "config-dir", 'C', "path", 0, "Directory to look for config files", 0 },
+    { 0 }
+};
+
+static GglBuffer config_path = GGL_STR("/etc/greengrass/config.yaml");
+static GglBuffer config_dir = GGL_STR("/etc/greengrass/config.d");
+
+static error_t arg_parser(int key, char *arg, struct argp_state *state) {
+    (void) arg;
+    (void) state;
+    switch (key) {
+    case 'c':
+        config_path = ggl_buffer_from_null_term(arg);
+        break;
+    case 'C':
+        config_dir = ggl_buffer_from_null_term(arg);
+        break;
+    case ARGP_KEY_END:
+        break;
+    default:
+        return ARGP_ERR_UNKNOWN;
+    }
+    return 0;
+}
+
+static struct argp argp = { opts, arg_parser, 0, doc, 0, 0, 0 };
 
 static void exit_cleanup(void) {
     ggconfig_close();
 }
 
 int main(int argc, char **argv) {
-    (void) argc;
-    (void) argv;
+    // NOLINTNEXTLINE(concurrency-mt-unsafe)
+    argp_parse(&argp, argc, argv, 0, 0, NULL);
 
     atexit(exit_cleanup);
 
     ggconfig_open();
+
+    ggconfig_load_file(config_path);
+    ggconfig_load_dir(config_dir);
 
     ggconfigd_start_server();